requirements.txt
alembic==1.13.1
annotated-types==0.6.0
anyio==4.3.0
click==8.1.7
exceptiongroup==1.2.0
fastapi==0.110.1
FastAPI-SQLAlchemy==0.2.1
greenlet==3.0.3
h11==0.14.0
httptools==0.6.1
idna==3.6
Mako==1.3.2
MarkupSafe==2.1.5
psycopg2-binary==2.9.9
pydantic==2.6.4
pydantic_core==2.16.3
python-dotenv==1.0.1
PyYAML==6.0.1
sniffio==1.3.1
SQLAlchemy==2.0.29
starlette==0.37.2
typing_extensions==4.11.0
uvicorn==0.29.0
uvloop==0.19.0
watchfiles==0.21.0
websockets==12.0
